feat(RouterLink): emit events onTransitionComplete(route, replace), onTransitionError(route, replace)
The idea is to be able to react to route changes depending on whether
the current history has been replaced or pushed to. Of course, error
handling could also be done in a routing guard, but it is somehow
difficult to get "has been pushed" vs. "has been replaced" info after
routing completion.